3 laba

+++ src/main/java/flowershop/service/CustomerService.java
@@ -1,22 +1,32 @@
 package flowershop.service;
 
+import flowershop.components.CustomerHashMap;
 import flowershop.dto.CustomerDto;
 import flowershop.entity.Customer;
+import flowershop.entity.SearchKey;
 import flowershop.entity.ShoppingCart;
+import flowershop.enums.OrderStatus;
 import flowershop.exception.TransactionDemoException;
 import flowershop.mapper.CustomerMapper;
 import flowershop.repository.CustomerRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import lombok.extern.slf4j.Slf4j;
 
+import java.time.LocalDate;
 import java.util.List;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class CustomerService {
 
     private final CustomerRepository customerRepository;
+    private final CustomerHashMap hashMap;
 
     private Customer findEntityById(Long id) {
         return customerRepository.findById(id).orElse(null);
@@ -37,27 +47,29 @@ public class CustomerService {
     @Transactional
     public CustomerDto createTransactional(CustomerDto dto) {
 
-        Customer customer = CustomerMapper.toEntity(dto);
-        customer = customerRepository.save(customer);
+        Customer customer = customerRepository.save(CustomerMapper.toEntity(dto));
 
         ShoppingCart cart = new ShoppingCart();
         cart.setCustomer(customer);
         cart.setId(customer.getId());
 
         customer.setCart(cart);
+
+        hashMap.clear();
 
         return CustomerMapper.toDto(customerRepository.save(customer));
     }
 
     public CustomerDto createWithoutTransaction(CustomerDto dto) {
-        Customer customer = CustomerMapper.toEntity(dto);
-        customer = customerRepository.save(customer);
+        Customer customer = customerRepository.save(CustomerMapper.toEntity(dto));
 
         ShoppingCart cart = new ShoppingCart();
         cart.setCustomer(customer);
         cart.setId(customer.getId());
 
         customer.setCart(cart);
+
+        hashMap.clear();
 
         throw new TransactionDemoException("Тест: Ошибка БЕЗ @Transactional.");
     }
@@ -65,8 +77,7 @@ public class CustomerService {
 
     @Transactional
     public CustomerDto createWithTransaction(CustomerDto dto) {
-        Customer customer = CustomerMapper.toEntity(dto);
-        customer = customerRepository.save(customer);
+        Customer customer = customerRepository.save(CustomerMapper.toEntity(dto));
 
         ShoppingCart cart = new ShoppingCart();
         cart.setCustomer(customer);
@@ -74,7 +85,7 @@ public class CustomerService {
 
         customer.setCart(cart);
 
-
+        hashMap.clear();
 
         throw new TransactionDemoException("Тест: Ошибка С @Transactional.");
     }
@@ -83,8 +94,11 @@ public class CustomerService {
     public CustomerDto update(Long id, CustomerDto dto) {
         Customer customer = findEntityById(id);
 
+
         customer.setName(dto.getName());
         customer.setPhoneNumber(dto.getPhoneNumber());
+
+        hashMap.clear();
 
         return CustomerMapper.toDto(customerRepository.save(customer));
     }
@@ -93,7 +107,54 @@ public class CustomerService {
     public void delete(Long id) {
         Customer customer = findEntityById(id);
 
+        if (customer == null) {
+
+            return;
+        }
+
+        hashMap.clear();
         customerRepository.delete(customer);
 
+    }
+
+    public Page<CustomerDto> findByFlower(Long flowerId, LocalDate date, int page, int size) {
+
+        List<String> orderStatuses = List.of(OrderStatus.PROCESSING.name(), OrderStatus.ACCEPTED.name());
+        Pageable pageable = PageRequest.of(page, size);
+
+        SearchKey key = new SearchKey(flowerId, orderStatuses, date, page, size);
+
+        if (hashMap.containsKey(key)) {
+            log.info("Данные взяты из кеша ");
+            return hashMap.get(key);
+        }
+
+        Page<Customer> customers = customerRepository.findByFlower(flowerId, date, orderStatuses, pageable);
+        Page<CustomerDto> customersDto = customers.map(CustomerMapper::toDto);
+
+        hashMap.put(key, customersDto);
+
+        return customersDto;
+
+    }
+
+    public Page<CustomerDto> findByFlowerNative(Long flowerId, LocalDate date, int page, int size) {
+
+        Pageable pageable = PageRequest.of(page, size);
+        List<String> orderStatuses = List.of(OrderStatus.PROCESSING.name(), OrderStatus.ACCEPTED.name());
+
+
+        SearchKey key = new SearchKey(flowerId, orderStatuses, date, page, size);
+
+        if (hashMap.containsKey(key)) {
+            log.info("Данные взяты из кеша");
+            return hashMap.get(key);
+        }
+
+        Page<Customer> customers = customerRepository.findByFlowerNative(flowerId, date, orderStatuses, pageable);
+        Page<CustomerDto> customersDto = customers.map(CustomerMapper::toDto);
+
+        hashMap.put(key, customersDto);
+        return customersDto;
     }
 }
